Conflicts:
	src/main/java/com/dc/esb/servicegov/service/impl/OperationServiceImpl.java
	src/main/webapp/jsp/service/search.jsp

+++ src/main/java/com/dc/esb/servicegov/service/impl/SystemLogServiceImpl.java
@@ -57,7 +57,7 @@ public class SystemLogServiceImpl  extends AbstractBaseService<OperationLog,Stri
                 hql += " and o.optDate < '" + values.get("endDate")[0] + "' ";
             }
         }
-        long count = operationLogDAO.findUnique(hql);
+        long count = (Long)operationLogDAO.findUnique(hql);
         return count;
     }
     public List<OperationLog> getLogs(Map<String, String[]> values, Page page) {
